docTR 2/requirements.txt
torch
transformers
streamlit
chromadb
sentence-transformers
pdf2image
pillow
python-dotenv
requests
numpy
pandas
langgraph
bitsandbytes
pdfplumber
fpdf
latex
langgraph
google-generativeai 
google-api-python-client 
